+++ Disney/src/main/java/com/mdelafuente/alkemy/disney/service/impl/MovieServiceImpl.java
@@ -105,7 +105,11 @@ public class MovieServiceImpl implements MovieService{
 	
 	@Override
 	public void delete(Long id) {
-		this.movieRepository.deleteById(id);	
+		Optional<MovieEntity> entity = this.movieRepository.findById(id);
+		if (!entity.isPresent()) {
+			throw new ParamNotFound("Movie ID not valid");
+		}
+		this.movieRepository.deleteById(id);
 	}
 
 }
